Remove dubious and broken call to ParserOutput::setEditSection()
It's dubious that this did what was actually intended given that it was the
last thing to execute in the special page, so it couldn't affect very much.

In any case, even if this did have some impact, having edit section links
show up in search results is still better than a special page that fatals.

Change-Id: I871475b80e2a5d63bb7580575b105260c2db045b